buildin/requirements.txt
asgiref==3.5.2
coverage==6.5.0
Django==4.1.1
django-phone-field==1.8.1
factory-boy==3.2.1
Faker==15.3.4
gunicorn==20.1.0
numpy==1.23.5
psycopg2==2.9.4
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
